* Add regression tests for dav merge back Bad file descriptor bug   with svn-trunk. * Make Editor::Merge set default pool at the beginning of each   confirmed file session - ensure_open and add_file.
git-svn-id: file:///home/chmrr/work/bps/complete-history/svk/orig-svn/trunk@807 4ccb8a5b-8ae0-0310-9b02-86531366991d

+++ t/tree.pl
@@ -1,5 +1,9 @@
 #!/usr/bin/perl
+
+my $pid = $$;
+
 END {
+    return unless $$ == $pid;
     rm_test($_) for @TOCLEAN;
 }
 
@@ -38,6 +42,7 @@ use SVK::XD;
 
 our @TOCLEAN;
 END {
+    return unless $$ == $pid;
     $SIG{__WARN__} = sub { 1 };
     cleanup_test($_) for @TOCLEAN;
 }
@@ -383,6 +388,7 @@ sub replace_file {
 }
 
 END {
+    return unless $$ == $pid;
     unlink $_ for @unlink;
 }
